holy... just got the validation messages working again on the signup screen.....
